Add files via upload
i have a fortran code which reads joints of a truss from a  .txt file where the first column shows the number of the joint the second column shows the coordinate in x direction and the third column shows the the coordinate of the joint in the y direction. then the code read from another file  .txt  members each consisting of two joints  the first column shows the number of the member the second one shows the first joint and the third  row shows the second joint. then is calculated the4 stiffness matrix for each member after that is created the global stiffness matrix. in the next step some loads are read from a .txt file where the first row shows the position in the x  axis and the second row the magnitude, the loads are applied only in the horizontal members. now that i have the load in the joints of specific horizontal members how can i extend it to find normal forces